+++ moin2gitwiki/moin2markdown.py
@@ -3,9 +3,20 @@ from pathlib import Path
 from typing import Optional
 
 import attr
+from bs4 import BeautifulSoup
+from furl import furl
+from sh import pandoc
 
 from .fetch_cache import FetchCache
 from .wikiindex import MoinEditEntry
+
+
+def is_a_linemark_para(tag):
+    return (
+        tag.name == "p"
+        and tag.has_attr("class")
+        and re.match(r"line\\d+", tag["class"][0])
+    )
 
 
 @attr.s(kw_only=True, frozen=True, slots=True)
@@ -13,125 +24,123 @@ class Moin2Markdown:
     #
     # -- attributes
     fetch_cache: FetchCache = attr.ib()
+    url_prefix: furl = attr.ib()
     link_table: dict = attr.ib()
     ctx = attr.ib(repr=False)
     #
-    # -- regular expressions used
-    moin_macro_pattern = re.compile(
-        r"""
-                (?:\<\<|\[\[)                           # opening part
-                (?P<macroname>  [A-Za-z0-9]+        )   # macro name
-                \(                                      # opening parens
-                (?P<params>     [^\]]*              )   # parameters
-                \)                                      # closing parens
-                (?:\>\>|\]\])                           # closing part
-                """,
-        re.VERBOSE,
-    )
-    ignored_moin_macros = (
-        "AttachInfo",
-        "AttachList",
-        "DateTime",
-        "FormCheckbox",
-        "FormFooter",
-        "FormHeader",
-        "FormRadio",
-        "FormSelect",
-        "FormSubmit",
-        "FormText",
-        "FormTextarea",
-        "FormUpload",
-        "FullSearch",
-        "GetText",
-        "GrabIt",
-        "Icon",
-        "IncVar",
-        "Include",
-        "MailTo",
-        "Navigation",
-        "PageList",
-        "RandomPage",
-        "RandomQuote",
-        "StatsChart",
-        "TableOfContents",
-    )
+    # prebuilt sh setup
+    pandoc = pandoc.bake("-f", "html", "-t", "gfm")
+    #
+    # smiley mapping
+    smiley_map = {
+        "X-(": ":rage:",
+        ":(": ":confused:",
+        ";)": ":wink:",
+        ":-?": ":stuck_out_tongue:",
+        ":-(": ":frowning_face:",
+        ";-)": ":wink:",
+        "{X}": ":x:",
+        "{3}": ":three:",
+        ":D": ":grin:",
+        ":)": ":slightly_smiling_face:",
+        "/!\\": ":warning:",
+        ":\\": ":confounded:",
+        ":-)": ":smiley:",
+        "|-)": ":pensive:",
+        "{i}": ":information_source:",
+        "{*}": ":star:",
+        "<:(": ":mask:",
+        "B)": ":sunglasses:",
+        "<!>": ":warning:",
+        ">:>": ":imp:",
+        "B-)": ":nerd_face:",
+        "(./)": ":white_check_mark:",
+        "{1}": ":one:",
+        "{o}": ":star:",
+        ":o": ":anguished:",
+        ":))": ":rofl:",
+        "(!)": ":bulb:",
+        "|)": ":monocle_face:",
+        ":-))": ":rofl:",
+        "{OK}": ":ok:",
+        "{2}": ":two:",
+    }
 
     @classmethod
-    def create_translator(cls, ctx, cache_directory: Path, link_table: dict):
+    def create_translator(
+        cls,
+        ctx,
+        cache_directory: Path,
+        url_prefix: str,
+        link_table: dict,
+    ):
         #
         # Build a fetch cache
         fetch_cache = FetchCache.initialise_cache(
             cache_directory=cache_directory,
             ctx=ctx,
         )
-        return cls(fetch_cache=fetch_cache, link_table=link_table, ctx=ctx)
+        return cls(
+            fetch_cache=fetch_cache,
+            link_table=link_table,
+            url_prefix=furl(url_prefix),
+            ctx=ctx,
+        )
 
-    def retrieve_and_translate(self, revision: MoinEditEntry) -> Optional[bytes]:
+    def retrieve_and_translate(self, revision: MoinEditEntry) -> Optional[str]:
+        # check if this revision has any content...
         lines = revision.wiki_content()
         if lines is None:
             return None
         else:
-            lines = self.translate(revision, lines)
-            lines.append("")  # additional line to give trailing \n
-            return "\n".join(lines).encode("utf-8")
+            target = self.url_prefix.copy()
+            target /= revision.page_path_unescaped()
+            target.args["action"] = "recall"
+            target.args["rev"] = revision.page_revision
+            content = self.fetch_cache.fetch(target.url)
+            main_content = self.extract_content_section(content)
+            translated = self.translate(main_content)
+            return translated
 
-    def translate(self, revision: MoinEditEntry, lines: list) -> list:
-        lines = self.pre_process_lines(revision=revision, lines=lines)
-        lines = self.process_lines(revision=revision, lines=lines)
-        return lines
+    def extract_content_section(self, html: str) -> str:
+        soup = BeautifulSoup(html, "html.parser")
+        content = soup.find(id="content")
+        #
+        # now strip out excess rubbish - anchor spans
+        for tag in content.find_all(class_="anchor"):
+            tag.decompose()
+        #
+        # Remove dead <p class="line???"> with no closer
+        for tag in content.find_all(is_a_linemark_para):
+            tag.unwrap()
+        #
+        # now find all the links, and if within the wiki, rewrite
+        for tag in content.find_all("a"):
+            target = tag["href"]
+            if target:
+                url = self.url_prefix.copy().join(target)
+                if url.url.startswith(self.url_prefix.url):
+                    new_url = url.url[len(self.url_prefix.url) :]
+                    if new_url in self.link_table:
+                        tag["href"] = self.link_table[new_url]
+                    else:
+                        tag.unwrap()
+            #
+            # strip any class attributes on links - tend to upset the translator
+            if tag.has_attr("class"):
+                del tag["class"]
+        #
+        # now find all the images and see if they map to emojis
+        # MoinMoin puts the emoji code in the title, so will purely match on that
+        for tag in content.find_all("img"):
+            if tag.has_attr("title") and tag["title"] in self.smiley_map:
+                tag.replace_with(" " + self.smiley_map[tag["title"]] + " ")
 
-    def pre_process_lines(self, revision: MoinEditEntry, lines: list) -> list:
-        new_lines = []
-        for line in lines:
-            # handle macro inserts
-            match = self.moin_macro_pattern.search(line)
-            if match is not None:
-                new_lines.extend(
-                    self.process_macro(
-                        revision=revision,
-                        line=line,
-                        macro_name=match.group("macroname"),
-                        params=match.group("params"),
-                    ),
-                )
-            else:
-                new_lines.append(line)
-        return new_lines
+        return content.contents
 
-    def process_lines(self, revision: MoinEditEntry, lines: list) -> list:
-        new_lines = []
-        for line in lines:
-            if line.startswith("#"):
-                new_lines.append("")  # ignore it
-            elif line.startswith("="):
-                line = line.rstrip()
-                line = line.rstrip("=")
-                line = line.rstrip()
-                line = line.replace("=", "#")
-                new_lines.append(line)
-                new_lines.append("")
-            else:
-                line = line.replace("'''", "**")  # bold
-                line = line.replace("''", "*")  # italic
-                line = line.replace("[[BR]]", "<br/>")  # breaks
-                new_lines.append(line)
-        return new_lines
+    def translate(self, input: str) -> str:
+        translated = self.pandoc(_in=input)
+        return translated.stdout
 
-    def process_macro(
-        self,
-        revision: MoinEditEntry,
-        line: str,
-        macro_name: str,
-        params: str,
-    ):
-        if macro_name == "IncludeUrlContentWiki":
-            return self.process_include_url_content_wiki(revision, params=params)
-        elif macro_name in self.ignored_moin_macros:
-            line = self.moin_macro_pattern.sub("", line)
-        else:
-            self.ctx.logger.warning(f"Unknown macro '{macro_name}'")
-        return [line]
 
-    def process_include_url_content_wiki(self, revision: MoinEditEntry, params: str):
-        url = params.strip().replace("%s", revision.unescape(revision.page_name))
-        return self.fetch_cache.fetch(url)
+# end